Refactor main, move some files to targets/cdm8e rename cdm_asm to cocas
