+++ cocas/targets/cdm8e/MacroLexer.py
@@ -1,4 +1,4 @@
-# Generated from /home/ilya/work/cdm8e/ORiGinalASM/cdm_asm/Macro.g4 by ANTLR 4.10.1
+# Generated from /home/ilya/work/cdm8e/ORiGinalASM/cocas/Macro.g4 by ANTLR 4.10.1
 from antlr4 import *
 from io import StringIO
 import sys
